@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core'
 import { BrowserModule } from '@angular/platform-browser'
-import { HttpClientModule } from '@angular/common/http'
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http'
 import { RouterModule } from '@angular/router'
 import { FormsModule } from '@angular/forms'
 
@@ -13,35 +13,28 @@ import { FloodWarningDetailsComponent } from './flood-warning-details/flood-warn
 import { FloodAreaComponent } from './flood-area/flood-area.component'
 import { AboutThisSiteComponent } from './about-this-site/about-this-site.component'
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    HeaderComponent,
-    ThemeSwitcherComponent,
-    HomeComponent,
-    FloodWarningsComponent,
-    FloodAreaComponent,
-    FloodWarningDetailsComponent,
-    AboutThisSiteComponent
-  ],
-  imports: [
-    BrowserModule,
-    HttpClientModule,
-    RouterModule.forRoot([
-      { path: 'home', component: HomeComponent },
-      { path: 'flood-warnings', component: FloodWarningsComponent },
-      {
-        path: 'flood-warning-details/:id',
-        component: FloodWarningDetailsComponent
-      },
-      { path: 'flood-area', component: FloodAreaComponent },
-      { path: 'about-this-site', component: AboutThisSiteComponent },
-      { path: '', redirectTo: 'home', pathMatch: 'full' },
-      { path: '**', redirectTo: 'home', pathMatch: 'full' }
-    ]),
-    FormsModule
-  ],
-  providers: [],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent,
+        HeaderComponent,
+        ThemeSwitcherComponent,
+        HomeComponent,
+        FloodWarningsComponent,
+        FloodAreaComponent,
+        FloodWarningDetailsComponent,
+        AboutThisSiteComponent
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        RouterModule.forRoot([
+            { path: 'home', component: HomeComponent },
+            { path: 'flood-warnings', component: FloodWarningsComponent },
+            {
+                path: 'flood-warning-details/:id',
+                component: FloodWarningDetailsComponent
+            },
+            { path: 'flood-area', component: FloodAreaComponent },
+            { path: 'about-this-site', component: AboutThisSiteComponent },
+            { path: '', redirectTo: 'home', pathMatch: 'full' },
+            { path: '**', redirectTo: 'home', pathMatch: 'full' }
+        ]),
+        FormsModule], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class AppModule {}
